requirements/base.txt
asgiref==3.3.4
coreapi==2.3.3
dj-rest-auth==2.1.4
django-allauth==0.44.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
django_extensions==3.1.3
drf-spectacular==0.18.2
Markdown==3.3.4
pytz==2021.1
sqlparse==0.2.4
shortuuid==1.0.1
Pillow==8.2.0
PyYAML==5.4.1
whitenoise==5.3.0
